# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Demo Loans System",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api/{version}",
            "description": "Api for loans system",
            "variables": {
                "version": {
                    "enum": [
                        "v1"
                    ],
                    "default": "v1"
                }
            }
        }
    ],
    "paths": {
...
        "/loans/search": {
            "post": {
                "tags": [
                    "Loan"
                ],
                "summary": "Get all loans",
                "description": "Get all loans to list.",
                "operationId": "1d391dddb9e9207978165710088f2ad9",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "minPrice": {
                                        "description": "minPrice",
                                        "type": "integer",
                                        "example": "100"
                                    },
                                    "maxPrice": {
                                        "description": "maxPrice",
                                        "type": "integer",
                                        "example": "10000"
                                    },
                                    "page": {
                                        "description": "page",
                                        "type": "integer",
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/loans": {
            "post": {
                "tags": [
                    "Loan"
                ],
                "summary": "Create new loan",
                "description": "Create new loan",
                "operationId": "26018426d0b8b1e7b9128b5deccdbd8f",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/newloan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loan"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/loans/detail/{loanId}": {
            "get": {
                "tags": [
                    "Loan"
                ],
                "summary": "Get loan detail",
                "description": "Get loan detail given by its id",
                "operationId": "334d7630ee6701b4c76fddcdb58e49a4",
                "parameters": [
                    {
                        "name": "loanId",
                        "in": "path",
                        "description": "ID of loan need get info.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/loans/update/{loanId}": {
            "put": {
                "tags": [
                    "Loan"
                ],
                "summary": "Update loan by id.",
                "description": "Update loan by id.",
                "operationId": "123f42df08dfc65b0a8cfd85f80b9b2f",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    },
                    {
                        "name": "loanId",
                        "in": "path",
                        "description": "ID of loan need update.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updateloan"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loan"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/loans/approve": {
            "patch": {
                "tags": [
                    "Loan"
                ],
                "summary": "Approve loan",
                "description": "Approve loan",
                "operationId": "aaecaf290485456b52b0b5154f60a512",
                "parameters": [
                    {
                        "name": "X-Requested-With",
                        "in": "header",
                        "description": "Request Header",
                        "schema": {
                            "type": "string",
                            "default": "XMLHttpRequest"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/loanapprove"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/loanapprove"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "loan": {
                "title": "Loan Model",
                "properties": {
                    "amount": {
                        "schema": "Loan",
                        "description": "amount borrowed",
                        "type": "integer",
                        "example": "1000"
                    },
                    "amount_to_pay": {
                        "schema": "Loan",
                        "description": "amount to pay",
                        "type": "float",
                        "example": "1000"
                    },
                    "interest": {
                        "schema": "Loan",
                        "description": "interest",
                        "type": "integer",
                        "example": "10"
                    },
                    "duration": {
                        "schema": "Loan",
                        "description": "duration",
                        "type": "integer",
                        "example": "15"
                    },
                    "date_applied": {
                        "schema": "Loan",
                        "description": "date_applied",
                        "type": "date",
                        "example": "2021-10-16"
                    },
                    "date_loan_ends": {
                        "schema": "Loan",
                        "description": "date_loan_ends",
                        "type": "date",
                        "example": "2021-11-30"
                    }
                },
                "type": "object"
            },
            "newloan": {
                "title": "New loan",
                "allOf": [
                    {
                        "properties": {
                            "user_id": {
                                "description": "customer id",
                                "type": "integer",
                                "example": "1"
                            }
                        },
                        "type": "object"
                    },
                    {
                        "$ref": "#/components/schemas/loan"
                    }
                ]
            },
            "updateloan": {
                "title": "Update loan",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/loan"
                    }
                ]
            },
            "loanapprove": {
                "title": "Loan approve",
                "allOf": [
                    {
                        "properties": {
                            "loan_id": {
                                "description": "Loan id",
                                "type": "integer",
                                "example": "1"
                            },
                            "approved": {
                                "description": "true | false",
                                "type": "boolean",
                                "example": "true"
                            },
                            "approved_date": {
                                "description": "approved_date",
                                "type": "date",
                                "example": "2021-10-16"
                            },
                            "payment_date": {
                                "description": "payment_date",
                                "type": "date",
                                "example": "2021-11-16"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "description": "Header value: `Bearer {access_token}`",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}